+++ lib/go_sell_sdk_flutter.dart
@@ -9,21 +9,16 @@ class GoSellSdkFlutter {
   static const String ERROR_CODE_INVALID_APP_CONFIGURATION = "501";
   static Map<dynamic, dynamic> _tapSDKResult = Map<dynamic, dynamic>();
 
-  static const MethodChannel _channel =
-      const MethodChannel('go_sell_sdk_flutter');
+  static const MethodChannel _channel = const MethodChannel('go_sell_sdk_flutter');
 
   static Future<dynamic> get startPaymentSDK async {
     if (!_validateSessionArge() || _validateAppConfig()) return _tapSDKResult;
 
     /// prepare sdk configurations
-    sdkConfigurations = {
-      "appCredentials": appCredentials,
-      "sessionParameters": sessionParameters
-    };
+    sdkConfigurations = {"appCredentials": appCredentials, "sessionParameters": sessionParameters};
 
     // /forward call to channel
-    dynamic result =
-        await _channel.invokeMethod('start_sdk', sdkConfigurations);
+    dynamic result = await _channel.invokeMethod('start_sdk', sdkConfigurations);
     print('result in dart : $result');
     return result;
   }
@@ -33,11 +28,7 @@ class GoSellSdkFlutter {
   static Map<String, dynamic> sessionParameters;
 
   /// App configurations
-  static void configureApp(
-      {String productionSecreteKey,
-      String sandBoxsecretKey,
-      String bundleId,
-      String lang}) {
+  static void configureApp({String productionSecreteKey, String sandBoxsecretKey, String bundleId, String lang}) {
     appCredentials = <String, dynamic>{
       "production_secrete_key": productionSecreteKey,
       "sandbox_secrete_key": sandBoxsecretKey,
@@ -71,6 +62,8 @@ class GoSellSdkFlutter {
     SDKMode sdkMode,
     PaymentType paymentType,
     bool allowsToSaveSameCardMoreThanOnce,
+    String cardHolderName,
+    bool allowsToEditCardHolderName,
   }) {
     sessionParameters = <String, dynamic>{
       'trxMode': trxMode.toString(),
@@ -96,18 +89,16 @@ class GoSellSdkFlutter {
       "SDKMode": sdkMode.toString(),
       "paymentType": paymentType.toString(),
       "allowsToSaveSameCardMoreThanOnce": allowsToSaveSameCardMoreThanOnce,
+      "cardHolderName": cardHolderName,
+      "editCardHolderName": allowsToEditCardHolderName,
     };
   }
 
   /// validate app configurations
   static bool _validateAppConfig() {
-    if (appCredentials["bundleId"] == "" ||
-        appCredentials["bundleId"] == "null" ||
-        appCredentials["bundleId"] == null) {
+    if (appCredentials["bundleId"] == "" || appCredentials["bundleId"] == "null" || appCredentials["bundleId"] == null) {
       _prepareConfigurationsErrorMap(
-          errorCode: ERROR_CODE_INVALID_APP_CONFIGURATION,
-          errorMsg: 'Invalid Bundle ID',
-          errorDescription: 'Bundle ID can not be empty or null');
+          errorCode: ERROR_CODE_INVALID_APP_CONFIGURATION, errorMsg: 'Invalid Bundle ID', errorDescription: 'Bundle ID can not be empty or null');
       return false;
     }
 
@@ -131,13 +122,9 @@ class GoSellSdkFlutter {
       return false;
     }
 
-    if (appCredentials["lang"] == "" ||
-        appCredentials["lang"] == "null" ||
-        appCredentials["lang"] == null) {
+    if (appCredentials["lang"] == "" || appCredentials["lang"] == "null" || appCredentials["lang"] == null) {
       _prepareConfigurationsErrorMap(
-          errorCode: ERROR_CODE_INVALID_SESSION_CONFIGURATION,
-          errorMsg: 'Invalid language',
-          errorDescription: 'Language can not empty or null');
+          errorCode: ERROR_CODE_INVALID_SESSION_CONFIGURATION, errorMsg: 'Invalid language', errorDescription: 'Language can not empty or null');
       return false;
     }
     return true;
@@ -145,9 +132,7 @@ class GoSellSdkFlutter {
 
   static bool _validateSessionArge() {
     // validate trx mode
-    if (sessionParameters["trxMode"] == "" ||
-        sessionParameters["trxMode"] == "null" ||
-        sessionParameters["trxMode"] == null) {
+    if (sessionParameters["trxMode"] == "" || sessionParameters["trxMode"] == "null" || sessionParameters["trxMode"] == null) {
       _prepareConfigurationsErrorMap(
           errorCode: ERROR_CODE_INVALID_SESSION_CONFIGURATION,
           errorMsg: 'Invalid Transaction Mode',
@@ -167,24 +152,18 @@ class GoSellSdkFlutter {
     }
 
     // validate customer
-    if (sessionParameters["customer"] == "null" ||
-        sessionParameters["customer"] == null) {
+    if (sessionParameters["customer"] == "null" || sessionParameters["customer"] == null) {
       _prepareConfigurationsErrorMap(
-          errorCode: ERROR_CODE_INVALID_SESSION_CONFIGURATION,
-          errorMsg: 'Invalid Customer',
-          errorDescription: 'Customer can not be empty or null');
+          errorCode: ERROR_CODE_INVALID_SESSION_CONFIGURATION, errorMsg: 'Invalid Customer', errorDescription: 'Customer can not be empty or null');
       return false;
     }
 
     // validate amount
-    if (sessionParameters["amount"] == "null" ||
-        !Util.isNumeric(sessionParameters["amount"]) ||
-        sessionParameters["amount"] == "") {
+    if (sessionParameters["amount"] == "null" || !Util.isNumeric(sessionParameters["amount"]) || sessionParameters["amount"] == "") {
       _prepareConfigurationsErrorMap(
           errorCode: ERROR_CODE_INVALID_SESSION_CONFIGURATION,
           errorMsg: 'Invalid Amount',
-          errorDescription:
-              'Amount can not be empty or null or zero and must be numeric value');
+          errorDescription: 'Amount can not be empty or null or zero and must be numeric value');
       return false;
     }
 
@@ -199,15 +178,12 @@ class GoSellSdkFlutter {
   }
 
   static void _validateBooleanValues(String param) {
-    if (sessionParameters[param] == "null" ||
-        sessionParameters[param] == null ||
-        sessionParameters[param] == "") {
+    if (sessionParameters[param] == "null" || sessionParameters[param] == null || sessionParameters[param] == "") {
       sessionParameters[param] = false;
     }
   }
 
-  static void _prepareConfigurationsErrorMap(
-      {String errorCode, String errorMsg, String errorDescription}) {
+  static void _prepareConfigurationsErrorMap({String errorCode, String errorMsg, String errorDescription}) {
     _tapSDKResult.putIfAbsent('sdk_result', () => 'SDK_ERROR');
     _tapSDKResult.putIfAbsent('sdk_error_code', () => errorCode);
     _tapSDKResult.putIfAbsent('sdk_error_message', () => errorMsg);